PR687: BasicService.getCodeBase() returns null for IcedTea6 1.9.7 + OSGI
The patch modifies how we try to find the JNLPClassLoader (from which we find
the ApplicationInstance). We first search the Context ClassLoader (and it's
parents) and then we search the ClassLoader for the classes on the stack (and
their parents). The Launcher always sets the Context ClassLoader of the
applications/applets it launches.