Reset the LogConfig in unit tests
LogConfig is globally shared. Fix unit tests that modify the logging
configuration to reset it after (and before) to avoid affecting other
unit tests or being affected by the results of other tests.

Before this patch, if the tests were executed in a certain order, it
was possible for them to leave behind a LogConfig that enabled logging
in the most verbose manner possible. This would generate megabytes of
output when running unit tests.

2013-10-22  Omair Majid  <omajid@redhat.com>

    * netx/net/sourceforge/jnlp/util/logging/LogConfig.java
    (resetLogConfig): New method.
    * tests/netx/unit/net/sourceforge/jnlp/util/logging/OutputControllerTest.java
    (setUp, tearDown): New method.