Change download progress indicator to a JDialog
The current download progress indicator window is a JFrame. Some
window managers in Linux assume it's an application window and try
to handle its size and position in a more appropriate way.  A tiling
window manager, for example, will resize it to fill the screen.

Change it to a JDialog so it's treated as a dialog and its size and
position are respected.

2014-05-02  Omair Majid  <omajid@redhat.com>

       * netx/net/sourceforge/jnlp/cache/DefaultDownloadIndicator.java: Switch to
       JDialog from JFrame. Rename frame to dialog and frameMutex to dialogMutex.
       Adjust all Callers.